matching2 - P7+P8: more public api additions, changes to context, corrections to multiple callback returns, corrections to A types, etc.
