+++ src/core/libraries/np/np_matching2/np_matching2.h
@@ -4,148 +4,13 @@
 #pragma once
 
 #include "common/types.h"
+#include "core/libraries/np/np_types2.h"
 
 namespace Core::Loader {
 class SymbolsResolver;
 }
 
 namespace Libraries::Np::NpMatching2 {
-
-using OrbisNpMatching2AttributeId = u16;
-using OrbisNpMatching2BlockKickFlag = u8;
-using OrbisNpMatching2ContextId = u16;
-using OrbisNpMatching2FlagAttr = u32;
-using OrbisNpMatching2Flags = u32;
-using OrbisNpMatching2LobbyId = u64;
-using OrbisNpMatching2LobbyMemberId = u16;
-using OrbisNpMatching2NatType = u8;
-using OrbisNpMatching2RequestId = u32;
-using OrbisNpMatching2RoomGroupId = u8;
-using OrbisNpMatching2RoomId = u64;
-using OrbisNpMatching2RoomMemberId = u16;
-using OrbisNpMatching2RoomPasswordSlotMask = u64;
-using OrbisNpMatching2ServerId = u16;
-using OrbisNpMatching2SignalingFlag = u8;
-using OrbisNpMatching2SignalingRequestId = u32;
-using OrbisNpMatching2TeamId = u8;
-using OrbisNpMatching2WorldId = u32;
-
-enum OrbisNpMatching2Event : u16 {
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_GET_WORLD_INFO_LIST = 0x0002,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_SET_ROOM_DATA_EXTERNAL = 0x0004,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_SET_USER_INFO = 0x0007,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_CREATE_JOIN_ROOM = 0x0101,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_JOIN_ROOM = 0x0102,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_LEAVE_ROOM = 0x0103,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_KICKOUT_ROOM_MEMBER = 0x0104,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_GRANT_ROOM_OWNER = 0x0105,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_SEARCH_ROOM = 0x0106,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_SEND_ROOM_MESSAGE = 0x0108,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_SET_ROOM_DATA_INTERNAL = 0x0109,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_GET_ROOM_DATA_INTERNAL = 0x010A,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_JOIN_LOBBY = 0x0201,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_SIGNALING_GET_PING_INFO = 0x0E01,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_MEMBER_JOINED = 0x1101,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_MEMBER_LEFT = 0x1102,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_KICKEDOUT = 0x1103,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_ROOM_DESTROYED = 0x1104,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_ROOM_OWNER_CHANGED = 0x1105,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_UPDATED_ROOM_DATA_INTERNAL = 0x1106,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_UPDATED_ROOM_MEMBER_DATA_INTERNAL = 0x1107,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_UPDATED_SIGNALING_OPT_PARAM = 0x1108,
-    ORBIS_NP_MATCHING2_SIGNALING_EVENT_DEAD = 0x5101,
-    ORBIS_NP_MATCHING2_SIGNALING_EVENT_ESTABLISHED = 0x5102,
-    ORBIS_NP_MATCHING2_SIGNALING_EVENT_NETINFO_ERROR = 0x5103,
-    ORBIS_NP_MATCHING2_CONTEXT_EVENT_START_OVER = 0x6F01,
-    ORBIS_NP_MATCHING2_CONTEXT_EVENT_STARTED = 0x6F02,
-    ORBIS_NP_MATCHING2_CONTEXT_EVENT_STOPPED = 0x6F03,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_CREATE_JOIN_ROOM_A = 0x7101,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_JOIN_ROOM_A = 0x7102,
-    ORBIS_NP_MATCHING2_REQUEST_EVENT_SEARCH_ROOM_A = 0x7106,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_MEMBER_JOINED_A = 0x8101,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_MEMBER_LEFT_A = 0x8102,
-    ORBIS_NP_MATCHING2_ROOM_EVENT_UPDATED_ROOM_MEMBER_DATA_INTERNAL_A = 0x8107,
-    ORBIS_NP_MATCHING2_ROOM_MSG_EVENT_MESSAGE_A = 0x9102,
-};
-
-enum OrbisNpMatching2EventCause : u8 {
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_LEAVE_ACTION = 1,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_KICKOUT_ACTION = 2,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_GRANT_OWNER_ACTION = 3,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_SERVER_OPERATION = 4,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_MEMBER_DISAPPEARED = 5,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_SERVER_INTERNAL = 6,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_CONNECTION_ERROR = 7,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_NP_SIGNED_OUT = 8,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_SYSTEM_ERROR = 9,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_CONTEXT_ERROR = 10,
-    ORBIS_NP_MATCHING2_EVENT_CAUSE_CONTEXT_ACTION = 11,
-};
-
-enum OrbisNpMatching2SignalingType : u8 {
-    ORBIS_NP_MATCHING2_SIGNALING_TYPE_NONE = 0,
-    ORBIS_NP_MATCHING2_SIGNALING_TYPE_MESH = 1,
-    ORBIS_NP_MATCHING2_SIGNALING_TYPE_STAR = 2,
-};
-
-enum OrbisNpMatching2SignalingConnectionInfoType : s32 {
-    ORBIS_NP_MATCHING2_SIGNALING_CONN_INFO_RTT = 1,
-    ORBIS_NP_MATCHING2_SIGNALING_CONN_INFO_BANDWIDTH = 2,
-    ORBIS_NP_MATCHING2_SIGNALING_CONN_INFO_PEER_NP_ID = 3,
-    ORBIS_NP_MATCHING2_SIGNALING_CONN_INFO_PEER_ADDR = 4,
-    ORBIS_NP_MATCHING2_SIGNALING_CONN_INFO_MAPPED_ADDR = 5,
-    ORBIS_NP_MATCHING2_SIGNALING_CONN_INFO_PACKET_LOSS = 6,
-    ORBIS_NP_MATCHING2_SIGNALING_CONN_INFO_PEER_NPID = 7,
-    ORBIS_NP_MATCHING2_SIGNALING_CONN_INFO_PEER_ADDRESS_A = 7,
-};
-
-enum OrbisNpMatching2AttrId : OrbisNpMatching2AttributeId {
-    ORBIS_NP_MATCHING2_ROOM_SEARCHABLE_INT_ATTR_EXTERNAL_1_ID = 0x004C,
-    ORBIS_NP_MATCHING2_ROOM_SEARCHABLE_INT_ATTR_EXTERNAL_2_ID = 0x004D,
-    ORBIS_NP_MATCHING2_ROOM_SEARCHABLE_INT_ATTR_EXTERNAL_3_ID = 0x004E,
-    ORBIS_NP_MATCHING2_ROOM_SEARCHABLE_INT_ATTR_EXTERNAL_4_ID = 0x004F,
-    ORBIS_NP_MATCHING2_ROOM_SEARCHABLE_INT_ATTR_EXTERNAL_5_ID = 0x0050,
-    ORBIS_NP_MATCHING2_ROOM_SEARCHABLE_INT_ATTR_EXTERNAL_6_ID = 0x0051,
-    ORBIS_NP_MATCHING2_ROOM_SEARCHABLE_INT_ATTR_EXTERNAL_7_ID = 0x0052,
-    ORBIS_NP_MATCHING2_ROOM_SEARCHABLE_INT_ATTR_EXTERNAL_8_ID = 0x0053,
-    ORBIS_NP_MATCHING2_ROOM_SEARCHABLE_BIN_ATTR_EXTERNAL_1_ID = 0x0054,
-    ORBIS_NP_MATCHING2_ROOM_BIN_ATTR_EXTERNAL_1_ID = 0x0055,
-    ORBIS_NP_MATCHING2_ROOM_BIN_ATTR_EXTERNAL_2_ID = 0x0056,
-    ORBIS_NP_MATCHING2_ROOM_BIN_ATTR_INTERNAL_1_ID = 0x0057,
-    ORBIS_NP_MATCHING2_ROOM_BIN_ATTR_INTERNAL_2_ID = 0x0058,
-    ORBIS_NP_MATCHING2_ROOMMEMBER_BIN_ATTR_INTERNAL_1_ID = 0x0059,
-};
-
-enum OrbisNpMatching2RoomFlagAttr : OrbisNpMatching2Flags {
-    ORBIS_NP_MATCHING2_ROOM_FLAG_ATTR_OWNER_AUTO_GRANT = 0x80000000,
-    ORBIS_NP_MATCHING2_ROOM_FLAG_ATTR_CLOSED = 0x40000000,
-    ORBIS_NP_MATCHING2_ROOM_FLAG_ATTR_FULL = 0x20000000,
-    ORBIS_NP_MATCHING2_ROOM_FLAG_ATTR_HIDDEN = 0x10000000,
-    ORBIS_NP_MATCHING2_ROOM_FLAG_ATTR_NAT_TYPE_RESTRICTION = 0x04000000,
-    ORBIS_NP_MATCHING2_ROOM_FLAG_ATTR_PROHIBITIVE_MODE = 0x02000000,
-};
-
-enum OrbisNpMatching2RoomMemberFlagAttr : OrbisNpMatching2Flags {
-    ORBIS_NP_MATCHING2_ROOMMEMBER_FLAG_ATTR_OWNER = 0x80000000,
-};
-
-enum OrbisNpMatching2Role : u8 {
-    ORBIS_NP_MATCHING2_ROLE_MEMBER = 1,
-    ORBIS_NP_MATCHING2_ROLE_OWNER = 2,
-};
-
-constexpr u32 ORBIS_NP_MATCHING2_SESSION_PASSWORD_SIZE = 8;
-constexpr u32 ORBIS_NP_MATCHING2_GROUP_LABEL_SIZE = 8;
-
-enum OrbisNpMatching2InitDefault : u64 {
-    ORBIS_NP_MATCHING2_THREAD_STACK_SIZE_DEFAULT = 32768,
-    ORBIS_NP_MATCHING2_POOLSIZE_DEFAULT = 131072,
-    ORBIS_NP_MATCHING2_SSL_POOLSIZE_DEFAULT = 196608,
-};
-
-struct OrbisNpMatching2InitializeParameter;
-struct OrbisNpMatching2CreateContextParameter;
-struct OrbisNpMatching2CreateContextParameterA;
 
 int PS4_SYSV_ABI sceNpMatching2Initialize(OrbisNpMatching2InitializeParameter* param);
 int PS4_SYSV_ABI sceNpMatching2Terminate();
