+++ src/core/libraries/np/np_matching2/np_matching2_internal.h
@@ -7,7 +7,6 @@
 #include <atomic>
 #include <chrono>
 #include <condition_variable>
-#include <deque>
 #include <map>
 #include <memory>
 #include <mutex>
@@ -17,12 +16,12 @@
 #include "common/types.h"
 #include "core/libraries/kernel/threads.h"
 #include "core/libraries/np/np_matching2/np_matching2.h"
-#include "core/libraries/np/np_matching2/np_matching2_types.h"
 #include "core/libraries/np/np_types.h"
 
 namespace shadnet {
 class CreateJoinRoomResponse;
 class GetWorldInfoListReply;
+class GetRoomDataExternalListReply;
 class LeaveRoomReply;
 class SearchRoomReply;
 } // namespace shadnet
@@ -96,6 +95,11 @@ struct CallbackPayload {
     std::unique_ptr<OrbisNpMatching2CreateJoinRoomResponseA> create_join_response_a;
     std::unique_ptr<OrbisNpMatching2LeaveRoomResponse> leave_room_response;
     std::unique_ptr<OrbisNpMatching2SearchRoomResponse> search_room_response;
+    std::unique_ptr<OrbisNpMatching2SearchRoomResponseA> search_room_response_a;
+    std::unique_ptr<OrbisNpMatching2GetRoomDataExternalListResponse>
+        room_data_external_list_response;
+    std::unique_ptr<OrbisNpMatching2GetRoomDataExternalListResponseA>
+        room_data_external_list_response_a;
     std::unique_ptr<OrbisNpMatching2GetWorldInfoListResponse> world_info_response;
     std::unique_ptr<OrbisNpMatching2SignalingGetPingInfoResponse> ping_info_response;
     std::vector<OrbisNpMatching2World> world_list;
@@ -107,14 +111,16 @@ struct CallbackPayload {
     std::vector<OrbisNpMatching2RoomBinAttrInternal*> room_bin_attr_ptrs;
     std::vector<OrbisNpMatching2RoomMemberBinAttrInternal> member_bin_attrs;
     std::vector<OrbisNpMatching2RoomDataExternal> room_data_external;
+    std::vector<OrbisNpMatching2RoomDataExternalA> room_data_external_a;
     std::vector<std::vector<u8>> bin_buffers;
-    std::deque<OrbisNpMatching2IntAttr> ext_int_attrs;
-    std::deque<OrbisNpMatching2BinAttr> ext_bin_attrs;
-    std::deque<OrbisNpMatching2RoomGroupInfo> ext_room_groups;
-    std::deque<Libraries::Np::OrbisNpId> ext_owner_npids;
+    std::vector<OrbisNpMatching2IntAttr> ext_int_attrs;
+    std::vector<OrbisNpMatching2BinAttr> ext_bin_attrs;
+    std::vector<OrbisNpMatching2RoomGroupInfo> ext_room_groups;
+    std::vector<Libraries::Np::OrbisNpId> ext_owner_npids;
     void* request_data = nullptr;
 
     std::unique_ptr<OrbisNpMatching2RoomMemberUpdate> room_member_update;
+    std::unique_ptr<OrbisNpMatching2RoomMemberUpdateA> room_member_update_a;
     std::unique_ptr<OrbisNpMatching2RoomUpdate> room_update;
     std::unique_ptr<OrbisNpMatching2RoomDataInternalUpdate> room_data_internal_update;
     std::unique_ptr<OrbisNpMatching2FlagAttr> event_chg_flag_attr;
@@ -122,6 +128,7 @@ struct CallbackPayload {
     std::unique_ptr<OrbisNpMatching2RoomPasswordSlotMask> event_chg_passwd_slot_mask;
     std::unique_ptr<OrbisNpMatching2RoomPasswordSlotMask> event_prev_passwd_slot_mask;
     std::unique_ptr<OrbisNpMatching2RoomMemberDataInternal> event_member;
+    std::unique_ptr<OrbisNpMatching2RoomMemberDataInternalA> event_member_a;
     void* room_event_data = nullptr;
 
     CallbackPayload() = default;
@@ -136,6 +143,9 @@ struct CallbackPayload {
         create_join_response_a.reset();
         leave_room_response.reset();
         search_room_response.reset();
+        search_room_response_a.reset();
+        room_data_external_list_response.reset();
+        room_data_external_list_response_a.reset();
         world_info_response.reset();
         ping_info_response.reset();
         world_list.clear();
@@ -147,6 +157,7 @@ struct CallbackPayload {
         room_bin_attr_ptrs.clear();
         member_bin_attrs.clear();
         room_data_external.clear();
+        room_data_external_a.clear();
         bin_buffers.clear();
         ext_int_attrs.clear();
         ext_bin_attrs.clear();
@@ -154,6 +165,7 @@ struct CallbackPayload {
         ext_owner_npids.clear();
         request_data = nullptr;
         room_member_update.reset();
+        room_member_update_a.reset();
         room_update.reset();
         room_data_internal_update.reset();
         event_chg_flag_attr.reset();
@@ -161,6 +173,7 @@ struct CallbackPayload {
         event_chg_passwd_slot_mask.reset();
         event_prev_passwd_slot_mask.reset();
         event_member.reset();
+        event_member_a.reset();
         room_event_data = nullptr;
     }
 };
@@ -168,6 +181,7 @@ struct CallbackPayload {
 struct ContextObject {
     OrbisNpMatching2ContextId ctx_id = 0;
     bool started = false;
+    bool a_variant = false;
 
     OrbisNpMatching2ServerId server_id = 1;
     OrbisNpServiceLabel service_label = 0;
@@ -214,6 +228,7 @@ struct ContextObject {
     void Reset() {
         ctx_id = 0;
         started = false;
+        a_variant = false;
         server_id = 1;
         service_label = 0;
         owner_np_id = {};
@@ -253,12 +268,12 @@ struct ContextObject {
 
 class ContextManager {
 public:
-    static constexpr u32 kMaxContexts = 255;
+    static constexpr u32 kMaxContexts = 10;
 
     static ContextManager& Instance();
 
     s32 CreateContext(const OrbisNpId* owner_np_id, OrbisNpServiceLabel service_label,
-                      OrbisNpMatching2ContextId* out_ctx_id);
+                      OrbisNpMatching2ContextId* out_ctx_id, bool a_variant = false);
 
     bool Check(OrbisNpMatching2ContextId ctx_id);
     ContextObject* Get(OrbisNpMatching2ContextId ctx_id);
@@ -351,6 +366,11 @@ void* BuildCreateJoinRoomPayloadA(ContextObject& ctx, const shadnet::CreateJoinR
 void* BuildLeaveRoomPayload(ContextObject& ctx, const shadnet::LeaveRoomReply& resp);
 void* BuildGetWorldInfoListPayload(ContextObject& ctx, const shadnet::GetWorldInfoListReply& resp);
 void* BuildSearchRoomPayload(ContextObject& ctx, const shadnet::SearchRoomReply& resp);
+void* BuildSearchRoomPayloadA(ContextObject& ctx, const shadnet::SearchRoomReply& resp);
+void* BuildGetRoomDataExternalListPayload(ContextObject& ctx,
+                                          const shadnet::GetRoomDataExternalListReply& resp);
+void* BuildGetRoomDataExternalListPayloadA(ContextObject& ctx,
+                                           const shadnet::GetRoomDataExternalListReply& resp);
 void* BuildGetRoomDataInternalPayload(ContextObject& ctx, OrbisNpMatching2RoomId room_id);
 
 void InitEventDispatcher();
