+++ src/core/libraries/np/np_matching2/np_matching2_mm.h
@@ -9,7 +9,6 @@
 
 #include "common/types.h"
 #include "core/libraries/np/np_matching2/np_matching2.h"
-#include "core/libraries/np/np_matching2/np_matching2_types.h"
 
 namespace ShadNet {
 class ShadNetClient;
@@ -31,6 +30,7 @@ enum class MmCommand : u16 {
     SetRoomDataExternal = 21,
     KickoutRoomMember = 22,
     GetWorldInfoList = 23,
+    GetRoomDataExternalList = 24,
 };
 
 void SetMmShadNetClient(std::shared_ptr<ShadNet::ShadNetClient> client,
@@ -61,7 +61,10 @@ s32 MmLeaveRoom(OrbisNpMatching2ContextId ctx_id, OrbisNpMatching2RequestId req_
 s32 MmGetWorldInfoList(OrbisNpMatching2ContextId ctx_id, OrbisNpMatching2RequestId req_id,
                        const OrbisNpMatching2GetWorldInfoListRequest& request);
 s32 MmSearchRoom(OrbisNpMatching2ContextId ctx_id, OrbisNpMatching2RequestId req_id,
-                 const OrbisNpMatching2SearchRoomRequest& request);
+                 const OrbisNpMatching2SearchRoomRequest& request, bool a_variant = false);
+s32 MmGetRoomDataExternalList(OrbisNpMatching2ContextId ctx_id, OrbisNpMatching2RequestId req_id,
+                              const OrbisNpMatching2GetRoomDataExternalListRequest& request,
+                              bool a_variant = false);
 s32 MmSetRoomDataInternal(OrbisNpMatching2ContextId ctx_id, OrbisNpMatching2RequestId req_id,
                           const OrbisNpMatching2SetRoomDataInternalRequest& request);
 s32 MmSetRoomDataExternal(OrbisNpMatching2ContextId ctx_id, OrbisNpMatching2RequestId req_id,
