+++ src/core/libraries/np/np_signaling/np_signaling_state.cpp
@@ -897,8 +897,9 @@ void StartHandshakeInitiator(OrbisNpSignalingConnectionId conn_id) {
     }
 }
 
-void QueueActivationLocked(OrbisNpSignalingConnectionId conn_id, std::string_view peer_online_id) {
-    g_pending_activations.push_back({conn_id, std::string(peer_online_id)});
+void QueueActivationLocked(OrbisNpSignalingConnectionId conn_id, std::string_view peer_online_id,
+                           bool start_handshake) {
+    g_pending_activations.push_back({conn_id, std::string(peer_online_id), start_handshake});
 }
 
 void ProcessPendingActivations() {
@@ -912,6 +913,7 @@ void ProcessPendingActivations() {
     }
 
     for (const PendingActivation& act : work) {
+        bool already_established = false;
         {
             SignalingMutexGuard lock;
             const auto it = g_connections.find(act.conn_id);
@@ -940,8 +942,16 @@ void ProcessPendingActivations() {
             it->second.addr = peer_addr;
             it->second.port = peer_port;
             SendActivationRequestLocked(it->second);
+            already_established = it->second.state == ConnState::Established;
+            LOG_INFO(Lib_NpSignaling, "connection {} sent local activation to '{}' at {:#x}:{}{}",
+                     act.conn_id, act.peer_online_id, peer_addr, sceNetNtohs(peer_port),
+                     act.start_handshake ? "" : " (existing connection)");
         }
-        StartHandshakeInitiator(act.conn_id);
+        if (act.start_handshake) {
+            StartHandshakeInitiator(act.conn_id);
+        } else if (already_established) {
+            EstablishConnection(act.conn_id, false);
+        }
     }
 }
 
@@ -1154,11 +1164,16 @@ void SendActivationRequestLocked(const ConnectionInfo& ci) {
 
 void SendControlCloseLocked(const ConnectionInfo& ci, ControlReason reason) {
     if (ci.addr == 0 || ci.port == 0) {
+        LOG_WARNING(Lib_NpSignaling,
+                    "Control Close SEND skipped: connId={} reason={} unresolved endpoint",
+                    ci.conn_id, static_cast<u16>(reason));
         return;
     }
     SignalingControl pkt = MakeControlLocked(ci, ControlKind::Close);
     pkt.reason = static_cast<u16>(reason);
-    Stubs::ControlSendTo(&pkt, sizeof(pkt), ci.addr, ci.port);
+    const int rc = Stubs::ControlSendTo(&pkt, sizeof(pkt), ci.addr, ci.port);
+    LOG_INFO(Lib_NpSignaling, "Control Close SEND: connId={} reason={} dst={:#x}:{} rc={}",
+             ci.conn_id, static_cast<u16>(reason), ci.addr, sceNetNtohs(ci.port), rc);
 }
 
 void HandleControlPacket(u32 from_addr, u16 from_port, const SignalingControl& pkt) {
@@ -1236,6 +1251,9 @@ void HandleControlPacket(u32 from_addr, u16 from_port, const SignalingControl& p
             ack_addr = ci.addr;
             ack_port = ci.port;
         } else if (conn_id == 0) {
+            LOG_INFO(Lib_NpSignaling,
+                     "Control RECV ignored: kind={} from='{}' src={:#x}:{} no matching connection",
+                     pkt.kind, from_id, from_addr, sceNetNtohs(from_port));
             return;
         } else if (kind == ControlKind::ActivationAck) {
             ConnectionInfo& ci = g_connections[conn_id];
@@ -1252,6 +1270,10 @@ void HandleControlPacket(u32 from_addr, u16 from_port, const SignalingControl& p
         } else if (kind == ControlKind::Close) {
             ConnectionInfo& ci = g_connections[conn_id];
             const ControlReason reason = static_cast<ControlReason>(pkt.reason);
+            LOG_INFO(Lib_NpSignaling,
+                     "Control Close RECV: connId={} from='{}' reason={} src={:#x}:{} state={}",
+                     conn_id, from_id, pkt.reason, from_addr, sceNetNtohs(from_port),
+                     static_cast<s32>(ci.state));
             if (reason == ControlReason::Deactivate && ci.state == ConnState::Established) {
                 peer_deactivated_conn = conn_id;
             } else {
@@ -1290,7 +1312,7 @@ void SendStunPing(s32 ctx_id) {
         online_id = it->second.owner_online_id;
     }
 
-    if (!Stubs::TransportIsReady()) {
+    if (!Stubs::EnsureTransport()) {
         return;
     }
 
@@ -1306,9 +1328,10 @@ void SendStunPing(s32 ctx_id) {
     StunPing ping{};
     ping.cmd = 0x01;
     std::memcpy(ping.online_id, online_id.data, ORBIS_NP_ONLINEID_MAX_LENGTH);
+    ping.local_ip = Stubs::AdvertisedAddr();
 
-    LOG_DEBUG(Lib_NpSignaling, "ctxId={} online_id='{}' server={:#x}:{}", ctx_id,
-              OnlineIdToString(online_id), server_addr, sceNetNtohs(server_udp));
+    LOG_DEBUG(Lib_NpSignaling, "ctxId={} online_id='{}' server={:#x}:{} local_ip={:#x}", ctx_id,
+              OnlineIdToString(online_id), server_addr, sceNetNtohs(server_udp), ping.local_ip);
 
     Stubs::SignalingSendTo(&ping, sizeof(ping), server_addr, server_udp);
 }
